feat: move zoom, sort setting from Navigator to Visualization icons, dropdown
